requirements/base.txt
Django==2.1
django-rest-auth==0.9.5
django-allauth==0.39.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
